fix type

+++ C0_Compiler/Shared.h
@@ -74,7 +74,7 @@ string exec(const char* cmd) {
 	std::string result;
 	std::unique_ptr<FILE, decltype(&_pclose)> pipe(_popen(cmd, "r"), _pclose);
 	if (!pipe) throw std::runtime_error("popen() failed!");
-	while (fgets(buffer.data(), buffer.size(), pipe.get()) != nullptr){
+	while (fgets(buffer.data(), (int)buffer.size(), pipe.get()) != nullptr){
 		result += buffer.data();
 	}
 	return result;
